moved ConeDims into cone_matrix_stuffing.py. Used this to bring various ConicSolvers apply functions into more of a common form. Wrote a big chunk of the ConicSolver contributing guide based on this standardization. Removed SuperSCS.
