+++ cvxpy/reductions/solvers/conic_solvers/__init__.py
@@ -24,6 +24,5 @@ from .glpk_mi_conif import GLPK_MI
 from .gurobi_conif import GUROBI
 from .mosek_conif import MOSEK
 from .scs_conif import SCS
-from .super_scs_conif import SuperSCS
 from .xpress_conif import XPRESS
 from .nag_conif import NAG
