+++ cvxpy/reductions/solvers/conic_solvers/conic_solver.py
@@ -29,61 +29,6 @@ import scipy.sparse as sp
 # performance regressions. If you are making a change to this file,
 # make sure to run cvxpy/tests/test_benchmarks.py to ensure that you have
 # not introduced a regression.
-class ConeDims(object):
-    """Summary of cone dimensions present in constraints.
-
-    Constraints must be formatted as dictionary that maps from
-    constraint type to a list of constraints of that type.
-
-    Attributes
-    ----------
-    zero : int
-        The dimension of the zero cone.
-    nonpos : int
-        The dimension of the non-positive cone.
-    exp : int
-        The dimension of the exponential cone.
-    soc : list of int
-        A list of the second-order cone dimensions.
-    psd : list of int
-        A list of the positive semidefinite cone dimensions, where the
-        dimension of the PSD cone of k by k matrices is k.
-    """
-    def __init__(self, constr_map):
-        self.zero = int(sum(c.size for c in constr_map[Zero]))
-        self.nonpos = int(sum(c.size for c in constr_map[NonPos]))
-        self.exp = int(sum(c.num_cones() for c in constr_map[ExpCone]))
-        self.soc = [int(dim) for c in constr_map[SOC] for dim in c.cone_sizes()]
-        self.psd = [int(c.shape[0]) for c in constr_map[PSD]]
-
-    def __repr__(self):
-        return "(zero: {0}, nonpos: {1}, exp: {2}, soc: {3}, psd: {4})".format(
-            self.zero, self.nonpos, self.exp, self.soc, self.psd)
-
-    def __str__(self):
-        """String representation.
-        """
-        return ("%i equalities, %i inequalities, %i exponential cones, \n"
-                "SOC constraints: %s, PSD constraints: %s.") % (self.zero,
-                                                                self.nonpos,
-                                                                self.exp,
-                                                                self.soc,
-                                                                self.psd)
-
-    def __getitem__(self, key):
-        if key == s.EQ_DIM:
-            return self.zero
-        elif key == s.LEQ_DIM:
-            return self.nonpos
-        elif key == s.EXP_DIM:
-            return self.exp
-        elif key == s.SOC_DIM:
-            return self.soc
-        elif key == s.PSD_DIM:
-            return self.psd
-        else:
-            raise KeyError(key)
-
 
 class LinearOperator(object):
     """A wrapper for linear operators."""
@@ -134,6 +79,8 @@ class ConicSolver(Solver):
     # Some solvers cannot solve problems that do not have constraints.
     # For such solvers, REQUIRES_CONSTR should be set to True.
     REQUIRES_CONSTR = False
+
+    EXP_CONE_ORDER = None
 
     def accepts(self, problem):
         return (isinstance(problem, ParamConeProg)
