+++ cvxpy/settings.py
@@ -57,7 +57,6 @@ CPLEX = "CPLEX"
 ECOS = "ECOS"
 SCS = "SCS"
 DIFFCP = "DIFFCP"
-SUPER_SCS = "SUPER_SCS"
 GUROBI = "GUROBI"
 OSQP = "OSQP"
 CPLEX = "CPLEX"
@@ -66,7 +65,7 @@ XPRESS = "XPRESS"
 NAG = "NAG"
 SOLVERS = [ECOS, CVXOPT, GLPK,
            GLPK_MI, SCS, GUROBI, OSQP, CPLEX,
-           MOSEK, CBC, XPRESS, SUPER_SCS, NAG]
+           MOSEK, CBC, XPRESS, NAG]
 
 # Xpress-specific items
 XPRESS_IIS = "XPRESS_IIS"
